Develop

+++ src/main/java/model/ProductBase.java
@@ -106,7 +106,7 @@ public abstract class ProductBase {
         }
     }
 
-    public float getSizeNumeric() {
+    public int getSizeNumeric() {
         return sizeNumeric.get();
     }
